Added allocation into bus

+++ hydrolib_bus/hydrolib_bus_datalink/include/hydrolib_bus_datalink_serializer.hpp
@@ -70,7 +70,7 @@ ReturnCode Serializer<TxStream, Logger>::Process(
   if (res < 0) {
     return ReturnCode::ERROR;
   }
-  if (res != current_message_.header.length) {
+  if (res != static_cast<int>(current_message_.header.length)) {
     return ReturnCode::OVERFLOW;
   }
   return ReturnCode::OK;
